object adaptation correction (#905)
